+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/BrowserConnected.java
@@ -28,4 +28,10 @@ public class BrowserConnected {
         portalGateway.incomingRequest(method, uri, new RequestOptions().setHeaders(headers), result::complete);
         return result.future().toCompletionStage();
     }
+
+    public CompletionStage<HttpClientResponse> request(HttpMethod method, String uri, MultiMap headers, String body) {
+        Promise<HttpClientResponse> result = Promise.promise();
+        portalGateway.incomingRequest(method, uri, new RequestOptions().setHeaders(headers), body, result::complete);
+        return result.future().toCompletionStage();
+    }
 }