+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add a default mapping from `.well-known/apple-app-site-association` to `base` to add support for Uniport mobile feature that require Apple's [associated domains](https://developer.apple.com/documentation/xcode/supporting-associated-domains).
 - Extracing helm values json schema from `portal-helm-chart-library` to automatically be applied to values files ([PORTAL-1467](https://issue.inventage.com/browse/PORTAL-1467)).
+- Support for back channel logout as triggered by Keycloak ([PORTAL-2175](https://issue.inventage.com/browse/PORTAL-2175)).
 
 ### Changed